BL-606  fix filtering on list pages and facets
Replaces the select-based filter with a more user-friendly and visually appealing custom filter component.

The new filter displays content type names with counts, sourced from facets provided by the API. It also persists filter selections in the URL query parameters, ensuring that selected filters are maintained across page reloads and navigation. This change also improves the display of the options by displaying the facet counts in the UI. Finally, the component disables types when the count is zero so that the user knows the selected type has no results.

Additionally, this commit fixes a server-side error logging issue by displaying the status text instead of the body, when there is a fetch response error.

Fixes the condition for determining if content has been migrated.