BL-731 Preserves query parameters on redirects-
causing nt7 widget link to view all fail in newtab
Ensures that query string parameters are maintained when redirecting to the homepage or a term alias. This prevents loss of potentially important URL parameters during locale-based redirects.

Also, it makes sure it fetches all aliases to be able to filter by langcode before redirecting to the locale alias

BL-731